+++ src/plugin/connector/block_storage/snapshot_connector.py
