+++ src/plugin/manager/__init__.py
@@ -1,8 +1,8 @@
 from plugin.manager.api_gateway.usage_plan_manager import UsagePlanManager
 from plugin.manager.api_gateway.api_key_manager import APIKeyManager
 from plugin.manager.api_gateway.service_manager import APIGatewayServiceManager
-from plugin.manager.block_storage.block_storage_type_manager import BlockStorageTypeManager
-from plugin.manager.block_storage.block_storage_snapshots_manager import BlockStorageSnapshotsManager
+from plugin.manager.block_storage.type_manager import BlockStorageTypeManager
+from plugin.manager.block_storage.snapshot_manager import BlockStorageSnapshotsManager
 from plugin.manager.block_storage.block_storage_manager import BlockStorageManager
 from plugin.manager.certificate_manager.certificate_manager import CertificateManager
 from plugin.manager.dns_plus.pool_manager import PoolManager
@@ -16,7 +16,7 @@ from plugin.manager.floating_ip.floating_ip_manager import FloatingIPManager
 from plugin.manager.image.image_manager import ImageManager
 from plugin.manager.instance.keypair_manager import KeyPairManager
 from plugin.manager.instance.instance_manager import InstanceManager
-from plugin.manager.instance.AZ_manager import InstanceAZManager
+from plugin.manager.instance.availability_zone_manager import InstanceAZManager
 from plugin.manager.loadbalancer.L7policies_manager import LoadBalancerL7PoliciesManager
 from plugin.manager.loadbalancer.member_manager import LoadBalancerMemberManager
 from plugin.manager.loadbalancer.ip_acl_manager import LoadBalancerIpACLManager
