+++ src/plugin/manager/instance/availability_zone_manager.py
@@ -2,7 +2,7 @@ import logging
 from spaceone.inventory.plugin.collector.lib import *
 
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
-from plugin.connector.instance.instance_AZ_connector import InstanceAZConnector
+from plugin.connector.instance.availability_zone_connector import InstanceAZConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
