+++ src/plugin/manager/security_group/rule_manager.py
@@ -2,7 +2,7 @@ import logging
 from spaceone.inventory.plugin.collector.lib import *
 
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
-from plugin.connector.sg.rules_connector import SGRulesConnector
+from plugin.connector.security_group.rule_connector import SGRulesConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
